+++ src/index.ts
@@ -10,7 +10,6 @@ export * from "./crypto/sea.js";
 export * from "./crypto/hmac.js";
 export * from "./crypto/derivation.js";
 export * from "./crypto/master-key.js";
-export * from "./crypto/webauthn.js";
 export * from "./server/challenge.js";
 export * from "./server/replay.js";
 export * from "./server/passport.js";